Refactor V1 api to be split up in multiple files plus expose decorator
